scroll the window to 0,0 when scroll is set to false
this scenario typically occurs when transitioning from a scrollable
mobile view to a fixed desktop wiew, eg:

(window :scroll (b true 720 false) ...)

without this feature, the window in this example would be frozen
at the position the user last scrolled when the window's width was
< 720.  there may be scenarios where this behaviour could be
undesireable; a developer might actually mant to freeze the window
in a particular position through this property, for example. it
might be worth considering a lock flag or alternative property
with the previous behavior when a real-world case manifests itself.